pyproject.toml
[project]
name = "robust-trinity"
version = "0.1.0"
description = "Structural Fault Tolerance in Heterogeneous Financial AI Systems under Adversarial Signal Injection"
requires-python = ">=3.11"
dependencies = [
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "anthropic>=0.39.0",
    "pydantic>=2.0",
    "wandb>=0.16.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "yfinance>=0.2.30",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "torch>=2.0",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
